split common code into separate folder
This is needed to run non OCP workloads (ex, ocm-api-load which tests
OCM server can be run as a standalone application on a jump host.)
using airflow. These workloads can use airflow common code to create
and run their dags. Next commit will have code for creating non-ocp
dag for ocm-api-load.
