+++ dags/openshift_nightlies/tasks/index/status.py
@@ -3,7 +3,7 @@ from os import environ
 
 from openshift_nightlies.util import var_loader, executor, constants
 from openshift_nightlies.models.release import OpenshiftRelease
-from openshift_nightlies.models.dag_config import DagConfig
+from common.models.dag_config import DagConfig
 
 import json
 import requests
