+++ dags/openshift_nightlies/tasks/install/hypershift/hypershift.py
@@ -4,7 +4,7 @@ from os import environ
 
 from openshift_nightlies.util import var_loader, kubeconfig, constants, executor
 from openshift_nightlies.tasks.install.openshift import AbstractOpenshiftInstaller
-from openshift_nightlies.models.dag_config import DagConfig
+from common.models.dag_config import DagConfig
 from openshift_nightlies.models.release import OpenshiftRelease
 from openshift_nightlies.tasks.benchmarks import e2e
