+++ dags/openshift_nightlies/tasks/utils/final_dag_status.py
@@ -1,7 +1,7 @@
 from os import environ
 from openshift_nightlies.util import var_loader, executor, constants
 from openshift_nightlies.models.release import OpenshiftRelease
-from openshift_nightlies.models.dag_config import DagConfig
+from common.models.dag_config import DagConfig
 
 
 from airflow.operators.python import PythonOperator
