+++ dags/openshift_nightlies/util/var_loader.py
@@ -1,5 +1,6 @@
 import json
 from os import environ
+from common.util import var_loader
 from openshift_nightlies.util import constants
 from openshift_nightlies.models.release import OpenshiftRelease
 
@@ -8,23 +9,9 @@ from airflow.models import Variable
 from kubernetes.client import models as k8s
 
 
-# Used to get the git user for the repo the dags live in. 
-def get_git_user():
-    git_repo = environ['GIT_REPO']
-    git_path = git_repo.split("https://github.com/")[1]
-    git_user = git_path.split('/')[0]
-    return git_user.lower()
-
-
-def get_secret(name, deserialize_json=False, required=True):
-    if required:
-        return Variable.get(name, deserialize_json=deserialize_json)
-    else:
-        try:
-            return Variable.get(name, deserialize_json=deserialize_json)
-        except KeyError:
-            return {}
-
+get_json=var_loader.get_json
+get_git_user=var_loader.get_git_user
+get_secret=var_loader.get_secret
 
 ### Task Variable Generator
 ### Grabs variables from appropriately placed JSON Files
@@ -56,13 +43,3 @@ def get_default_task_vars(release: OpenshiftRelease, task="install", task_dir=f"
     else:
         file_path = f"{task_dir}/{task}/defaults.json"
     return get_json(file_path)
-
-
-def get_json(file_path):
-    try: 
-        with open(file_path) as json_file:
-            return json.load(json_file)
-    except IOError as e:
-        return {}
-    except Exception as e:
-        raise Exception(f"json file {file_path} failed to parse")
