+++ dags/tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 import json
 
-from openshift_nightlies.models.dag_config import DagConfig
+from common.models.dag_config import DagConfig
 from openshift_nightlies.models.release import OpenshiftRelease, BaremetalRelease
